requirements.txt
numpy
requests
torch
torchvision
pillow
fastapi
pydantic
starlette
uvicorn
aiofiles
python-dotenv